[Reporting] Reorganize UI components
