+++ x-pack/plugins/reporting/public/management/report_diagnostic.tsx
