+++ x-pack/plugins/reporting/public/management/report_download_button.tsx
@@ -7,8 +7,8 @@
 
 import { EuiButtonIcon, EuiToolTip } from '@elastic/eui';
 import React, { FunctionComponent } from 'react';
-import { JOB_STATUSES } from '../../../common/constants';
-import { Job as ListingJob, Props as ListingProps } from '../report_listing';
+import { JOB_STATUSES } from '../../common/constants';
+import { Job as ListingJob, Props as ListingProps } from './report_listing';
 
 type Props = { record: ListingJob } & ListingProps;
