+++ x-pack/plugins/reporting/public/notifier/general_error.tsx
