+++ x-pack/plugins/reporting/public/notifier/index.ts
@@ -5,10 +5,8 @@
  * 2.0.
  */
 
-export { getSuccessToast } from './job_success';
 export { getFailureToast } from './job_failure';
+export { getGeneralErrorToast } from './general_error';
+export { getSuccessToast } from './job_success';
 export { getWarningFormulasToast } from './job_warning_formulas';
 export { getWarningMaxSizeToast } from './job_warning_max_size';
-export { getGeneralErrorToast } from './general_error';
-export { ScreenCapturePanelContent } from './screen_capture_panel_content';
-export { getSharedComponents } from './shared';
