+++ x-pack/plugins/reporting/public/notifier/job_completion_notifications.ts
