+++ x-pack/plugins/reporting/public/notifier/job_failure.tsx
