+++ x-pack/plugins/reporting/public/share_context_menu/panel_spinner.tsx
