+++ x-pack/plugins/reporting/public/share_context_menu/register_csv_reporting.tsx
@@ -16,9 +16,9 @@ import type { ShareContext } from '../../../../../src/plugins/share/public';
 import type { LicensingPluginSetup } from '../../../licensing/public';
 import { CSV_JOB_TYPE } from '../../common/constants';
 import type { JobParamsCSV } from '../../server/export_types/csv_searchsource/types';
-import { ReportingPanelContent } from '../components/reporting_panel_content_lazy';
 import { checkLicense } from '../lib/license_check';
 import type { ReportingAPIClient } from '../lib/reporting_api_client';
+import { ReportingPanelContent } from './reporting_panel_content_lazy';
 
 export const ReportingCsvShareProvider = ({
   apiClient,
