+++ x-pack/plugins/reporting/public/share_context_menu/reporting_panel_content.tsx
