+++ x-pack/plugins/reporting/public/share_context_menu/screen_capture_panel_content.tsx
