+++ x-pack/plugins/reporting/public/shared/index.ts
